requirements-cpu.txt
# CPU 环境版本 - 适用于 Mac 和没有 GPU 的机器
# --------- TensorFlow --------- #
tensorflow>=2.13.0              # 通用 TensorFlow 版本
# 对于 M 系列芯片的 Mac 用户，如需优化可取消下面两行注释
# tensorflow-macos>=2.13.0      # Mac 专用 TensorFlow 版本
# tensorflow-metal>=1.0.0       # 针对 Apple M 系列芯片的 Metal 加速

# --------- JAX (CPU) --------- #
jax[cpu]>=0.4.13                # JAX CPU 版本
flax>=0.7.0                     # JAX 的神经网络库
optax>=0.1.7                    # JAX 的优化库

# --------- Hydra --------- #
hydra-core==1.3.2               # Hydra 核心库
hydra-colorlog==1.2.0           # 彩色日志输出
hydra-optuna-sweeper==1.2.0     # 使用 Optuna 进行超参数搜索的 Sweeper

# --------- 其他 --------- #
rootutils       # 标准化项目根目录设置
pre-commit      # 提交代码时使用 lint 之类的钩子
rich            # 在终端中实现更美观的文本输出
pytest          # 单元测试
pyDOE           # 实验设计
scipy           # 科学计算
matplotlib      # 可视化
tqdm            # 进度条
